requirements.txt
django == 1.11.3
psycopg2 == 2.7.1
djangorestframework == 3.6.2
requests == 2.13.0
django-cors-headers == 2.1.0
boto == 2.48.0
django-storages == 1.6.3
gunicorn == 19.7.1
django-widget-tweaks==1.4.1